# Conflicts:
#	app/Exception/ErrorHttpException.php

+++ app/Middleware/CheckAuthMiddleware.php
@@ -27,7 +27,8 @@ class CheckAuthMiddleware extends MiddlewareAbstract
 		if (empty($user)) {
 			throw new ErrorHttpException('请先登录', [], 444);
 		}
-		if (!icache()->has(sprintf(UserLogic::USER_LOGOUT_AFTER_CHANGE_PWD, $user['uid']))) {
+		//如果修改密码后强制退出
+		if (empty($user['login_from_app']) && !icache()->has(sprintf(UserLogic::USER_LOGOUT_AFTER_CHANGE_PWD, $user['uid']))) {
 			$request->session->destroy();
 			throw new ErrorHttpException('请先登录', [], 444);
 		}
